requirements.txt
flask==3.1.*
flask-cors==5.*
groq==0.13.*
python-dotenv==1.*
sentence-transformers>=3.0.0
numpy>=1.24.0
psycopg[binary]>=3.2.0
